feat(meta): fix test error
