+++ engine/pkg/orm/mock.go
@@ -32,33 +32,6 @@ func randomDBFile() string {
 	return uuid.NewGenerator().NewString() + ".db"
 }
 
-// NewMockEpochClient is the mock for EpochClient
-func NewMockEpochClient() model.EpochClient {
-	return &mockEpochClient{
-		epoch: 1,
-	}
-}
-
-type mockEpochClient struct {
-	sync.Mutex
-	epoch int64
-}
-
-func (m *mockEpochClient) Initialize(ctx context.Context) error {
-	return nil
-}
-
-func (m *mockEpochClient) Close() error {
-	return nil
-}
-
-func (m *mockEpochClient) GenEpoch(ctx context.Context) (int64, error) {
-	m.Lock()
-	defer m.Unlock()
-	m.epoch = m.epoch + 1
-	return m.epoch, nil
-}
-
 // NewMockClient creates a mock orm client
 func NewMockClient() (Client, error) {
 	// ref:https://www.sqlite.org/inmemorydb.html
@@ -88,4 +61,31 @@ func NewMockClient() (Client, error) {
 	}
 
 	return cli, nil
+}
+
+// NewMockEpochClient is the mock for EpochClient
+func NewMockEpochClient() model.EpochClient {
+	return &mockEpochClient{
+		epoch: 1,
+	}
+}
+
+type mockEpochClient struct {
+	sync.Mutex
+	epoch int64
+}
+
+func (m *mockEpochClient) Initialize(ctx context.Context) error {
+	return nil
+}
+
+func (m *mockEpochClient) Close() error {
+	return nil
+}
+
+func (m *mockEpochClient) GenEpoch(ctx context.Context) (int64, error) {
+	m.Lock()
+	defer m.Unlock()
+	m.epoch = m.epoch + 1
+	return m.epoch, nil
 }
